Fix issue with bad completions when backspacing
By default, Vim chooses to perform its own filtering of the items in the
completion menu when the user presses backspace or C-H if the menu is
already shown. We need to disabuse Vim of the notion that it knows what it's
doing and to let us handle it.

Fixes #526.