Made changes to the install.sh to recognize .dll.a files on CYGWIN allowing it to discover PYTHON_INCLUDE_DIR and PYTHON_LIBRARY
Also changed wrap_python.hpp to only check for defined(__CYGWIN__)
on 32-bit systems. Previously this check was OR'd with the check
for defined(__WIN32__) causing the system to erroneously update the
SIZEOF_LONG value and subsequently resulting in "make ycm_core"
failing with "LONG_BIT definition appears wrong for platform" on
CYGWIN-64 systems.