add g:ycm_diagnostic_ignore_files option
Use this option to specify files that do not display diagnostic UI. It's
a list of regular-expression patterns. The full paths of files are
matched against these patterns, and the matches are case sensitive.

This option is part of the Syntastic compatibility layer; if the option
is not set, YCM will fall back to the value of the
`g:syntastic_ignore_files` option before using this option's default.