Use python2 in install.sh if it exists
This checks if there exists a python2 in $PATH and uses this instead of python to call the build script. This fixes the build for archlinux where python is python2 and also works on systems that don't have a python2 binary (in which case python is python2). This check should work in all POSIX shells.