Ignore dirty submodules in status
Added flag to remove dirty submodules to come up in git status lines.
This allows automated tools, such as ansible, which rely on git status
output, to correctly identify whether YCM needs to be recompiled or not.
Otherwise, the status after compilation is always "modified" for the
submodules, leading to incorrect understanding by said tools of the
status of the repository.