Catch and handle respones when files are already open
For multi-file refactors/FixIts, the file may be open in other editor or
it may just have a swap file. When this happens, we handle the response
when the response is Abort or Quit. We simply raise an exception and rely
on the user to manually fix their files using the undo command.