Improve experience for opening new files
Previously opening a number of new splits and the location list window would
almost always lead to the original window containing the user's cursor being
tiny. This is due to the way splits are opening in Vim by default.

The improvement is to use 'previewheight' and fixed height for the new splits.
This helps ensure that the original window is still visible and the user is not
unduly disorientated.

Finally, we issue a note to the user that we're about to open a bunch of new
splits, giving her the option to abort cleanly. This isn't strictly required,
but it is possible for this operation to hit a large number of files, so we
don't want that to just appear to spam all over the user's screen. The UX when
opening this many files wouldn't be great anyway, but a warning seems the
simplest way to signpost it.