Close the splits after editing
This leaves them as hidden buffers, allowing use of the quickfix list
(with ctrl-w <enter>) to open the files in splits.

In order to prevent this being very slow and/or leaving strange artefacts
(like location lists), we disable autocommands when opening.

The drawback of this is the buffers are not set up correctly when edited,
i.e. BufRead autocommands etc. which do filetype detection are not run.
This is slightly bad, but not as bad as the UX when opening many files
being extrmely slow and leaving unexpected windows on the screen.