Make it more clear that install.py require python2
Installation won't actually fail when running with python3. It will just result in YCM not working.

I just wasted half an hour on this. This is my attempt to make sure no1 else has the same error.

I finally figured out the YCM server wouldn't run with this error:
`ImportError: dynamic module does not define init function (initycm_core)`

Which was caused by running the installation from python3. Which is stated only in this issue:
https://github.com/Valloric/YouCompleteMe/issues/6