Provide config to disable autoload
Even deferring the autoload to the `VimEnter` stage isn't enough to prevent noticeable slowdown in booting vim. Executing `youcompleteme#Enable()` can still take hundreds of milliseconds even on a modern piece of hardware.

This allows users to disable the default autoloading if they wish so that they can execute `youcompleteme#Enable()` at a more appropriate time and get to an editable buffer faster.