add caveat for non-Vundle installs
I switched from Vundle to vim-plug a while back, but when I was following the steps for the install again, I didn't think twice about the location of my packages. Since I actually still had YouCompleteMe in my `~/.vim/bundle/` directory, nothing seemed off, and I wasted a bit of time until I realized that I was just in the wrong directory.

Given that vim-plug is becoming a fairly popular package manager, I figured adding a note to the documentation might save future folks the trouble. In terms of where to add it, putting it in each OS type seemed like the most obvious remedy, but other suggestions are welcome too.