Add g:ycm_start_autocmd option
Profiling with `vim --startuptime` shows that YCM is responsible for
about 250ms of the overall 500ms that Vim spends starting up on my
machine. By switching from "VimEnter" to "CursorHold,CursorHoldI" as a
trigger for calling `youcompleteme#Enable`, we get almost all of that
back, leading to an approximately 50% speed-up and a very perceptible
improvement in responsiveness.

The trade-off here is that YCM won't get fully initialized until Vim has
been idle for `'updatetime'` seconds (2s in my set-up, which apparently
YCM itself sets; before YCM runs though, the default of 4s is probably
active). As not everybody may like this trade-off, I've made all this
configurable via an option.