Reduced latency by using jobstart to dispatch request (optional)
Handling of completions is already asynchronous but dispatching the
request is not. As this is done on every key stroke it does influence
input latency for most vim flavors. The option
g:ycm_async_completionrequest_dispatch enables an asynchronous
dispatching by using the jobstart (or job_start).
It uses "true" as there is no real job to be done in the shell, but we
only want to trigger the callback.