Support for signature help
This uses Vim's new popup windows (where available). When the functions
we require are not available, we simply don't send the request and thus
never attempt to show any signatures.

The biggest challenge is placement of the popup, which is essentially
manual for us. Note that the popup position passed to Vim must be the
_screen_ position, not a buffer position. So we attempt to mimic
the way text properties follow text, by anchoring the popup to a
specific character in a specific buffer. We use the 'screenpos' function
for this.

The idea is to place the popup:

* imemdiately above the triggering character (named the 'anchor'), not
  overlapping the cursor if there is room above for all the items.
* otherwise, immediately below the triggering character.

Complications include:

* Ensuring there is room. If there is no room, just don't display.
* Avoiding overlapping the PUM. This is tricky as the pum is displayed
  asynchronously. We use the CompleteChanged event to try and recaculate
  our popup position when it moves. Again, if we can't display without
  overlapping, just don't display.

The rationale for not displaying is that signature help is secondary
information that may also be abailabel from the preview window (or
preview popup), so should not interfere with the primary completion
mechanism.