Read default settings to populate undocumented variables
This means we don't have to have hard-coded list of known options in the
file. There are a number of ycmd options which we don't document,
because they are added only as 'advanced' features (mainly used by the
ycmd devs). It makes little sense to have to hard-code them in every
client, so we load the default settings from the json.