Add vimscript testing framework based on vim's own test framework
The included README explains most of how the framework works.

The basic idea is that we run Vim with
`-S run_test.vim a_test.test.vim`.  This sources a_test.test.vim and
runs all the functions in it named Test_something. Tests use assert*
functions to push errors into v:errors and the run_tests.vim script
handles things like:

* saving output
* saving ycm/ycmd log files when a failure is detected
* catching exceptions, and unexpected vim exits
* a bit of tear down and clean up

There's a script to run the test and a makefile to do the same
(as well as clean up the log spam).

Finally, there's some azure infrastructure to run the tests in the CI
environment. We use an independent job (on Linux only) on the basis that
it's easy to do and some testing is better than no testing.

Coverage testing is also supported for the python parts of the YCM
client code running inside Vim. Unfortunately this requires using a
submodule to contain the coveragepy implementation. I'd like to not do
this becuase it requires all users to install this, but getting the
module loaded inside vim is not simple. An alternative might be to
simply clone it manually when running the tests.