Do not hardcode async diagnostic filetype support
Hardcoding supported filetypes prooved to be error prone, as we have
forgot to add `go`, `rust` and in my fork `swift` to the set. Another
problem with hardcoded filetypes is the `GenericLSPCompleter` which can
provide diagnostics for arbitrary filetypes.

This PR does away with `DIAGNOSTIC_UI_ASYNC_FILETYPES` and
`DIAGNOSTIC_UI_FILETYPES`.