Signature Help: Enable syntax highlighting in the popup
We *don't* want to enable 'filetype' in the popup - this actually leads
to YCM trying to parse the signature help buffer and various other
autocommands the we don't want. Just enabling the syntax allows for
Syntax autocommnds to even customise the display (I guess), but doesn't
have lots of lag like filetype setting.

In order to differentiate the current signature, we used to use
some text properties to mimic the completion popup. This doesn't work
with syntax highlighting (it overrides it), so we remove that and use
the syntax file's 'CursorLine' highlighting. This works nicely.

Note: There's a "cursorline" option to popup_open, but we don't want to
use that, as it actually uses 'PMenuSel', and is essentially designed
for poup menu-like things. We actually use PMenuSel for the current
_argument_ highlighting anyway.